refactor: extract telegram pending operation handler
